+++ api/src/Dom/Port/OccasionPort.php
@@ -4,10 +4,13 @@ declare(strict_types=1);
 
 namespace App\Dom\Port;
 
+use App\Dom\Exception\OccasionPasseeException;
 use App\Dom\Exception\PasAdminException;
 use App\Dom\Exception\PasParticipantException;
 use App\Dom\Exception\PasParticipantNiAdminException;
 use App\Dom\Exception\PasUtilisateurNiAdminException;
+use App\Dom\Exception\TirageDejaLanceException;
+use App\Dom\Exception\TirageEchoueException;
 use App\Dom\Exception\UtilisateurDejaParticipantException;
 use App\Dom\Exception\UtilisateurOffreOuRecoitDejaException;
 use App\Dom\Model\Auth;
@@ -15,9 +18,9 @@ use App\Dom\Model\Occasion;
 use App\Dom\Model\Resultat;
 use App\Dom\Model\Utilisateur;
 use App\Dom\Plugin\MailPlugin;
+use App\Dom\Repository\ExclusionRepository;
 use App\Dom\Repository\OccasionRepository;
 use App\Dom\Repository\ResultatRepository;
-use App\Dom\Repository\UtilisateurRepository;
 use App\Infra\Tools\ArrayTools;
 use DateTime;
 use Doctrine\DBAL\Exception\UniqueConstraintViolationException;
@@ -25,20 +28,20 @@ use Doctrine\DBAL\Exception\UniqueConstraintViolationException;
 class OccasionPort
 {
     private $mailPlugin;
+    private $exclusionRepository;
     private $occasionRepository;
     private $resultatRepository;
 
     public function __construct(
         MailPlugin $mailPlugin,
+        ExclusionRepository $exclusionRepository,
         OccasionRepository $occasionRepository,
-        ResultatRepository $resultatRepository,
-        UtilisateurRepository $utilisateurRepository
-    )
-    {
+        ResultatRepository $resultatRepository
+    ) {
         $this->mailPlugin = $mailPlugin;
+        $this->exclusionRepository = $exclusionRepository;
         $this->occasionRepository = $occasionRepository;
         $this->resultatRepository = $resultatRepository;
-        $this->utilisateurRepository = $utilisateurRepository;
     }
 
     /**
@@ -49,16 +52,14 @@ class OccasionPort
         Auth $auth,
         Occasion $occasion,
         Utilisateur $participant
-    ): Occasion
-    {
+    ): Occasion {
         if (!$auth->estAdmin()) throw new PasAdminException();
 
         $occasion->addParticipant($participant);
 
         try {
             $occasion = $this->occasionRepository->update($occasion);
-        }
-        catch (UniqueConstraintViolationException $err) {
+        } catch (UniqueConstraintViolationException $err) {
             throw new UtilisateurDejaParticipantException();
         }
 
@@ -79,10 +80,9 @@ class OccasionPort
         Occasion $occasion,
         Utilisateur $quiOffre,
         Utilisateur $quiRecoit
-    ): Resultat
-    {
+    ): Resultat {
         if (!$auth->estAdmin()) throw new PasAdminException();
-        
+
         $participants = $occasion->getParticipants();
         if (
             !ArrayTools::some($participants, [$quiOffre, 'estUtilisateur']) ||
@@ -93,8 +93,7 @@ class OccasionPort
 
         try {
             $resultat = $this->resultatRepository->create($occasion, $quiOffre, $quiRecoit);
-        }
-        catch (UniqueConstraintViolationException $err) {
+        } catch (UniqueConstraintViolationException $err) {
             throw new UtilisateurOffreOuRecoitDejaException();
         }
 
@@ -112,8 +111,7 @@ class OccasionPort
         Auth $auth,
         DateTime $date,
         string $titre
-    ): Occasion
-    {
+    ): Occasion {
         if (!$auth->estAdmin()) throw new PasAdminException();
 
         $occasion = $this->occasionRepository->create(
@@ -132,8 +130,7 @@ class OccasionPort
         Auth $auth,
         Occasion $occasion,
         array &$resultats = null
-    ): Occasion
-    {
+    ): Occasion {
         if (
             !$auth->estAdmin() &&
             !ArrayTools::some($occasion->getParticipants(), [$auth, 'estUtilisateur'])
@@ -150,13 +147,99 @@ class OccasionPort
     }
 
     /**
+     * @throws PasAdminException
+     * @throws TirageDejaLanceException
+     * @throws TirageEchoueException
+     * @return Resultat[]
+     */
+    public function lanceTirage(
+        Auth $auth,
+        Occasion $occasion,
+        bool $force,
+        ?int $nbMaxIter
+    ): array {
+        if (!$auth->estAdmin()) throw new PasAdminException();
+
+        if ($occasion->getDate() <= new DateTime()) throw new OccasionPasseeException();
+
+        if ($this->resultatRepository->hasForOccasion($occasion)) {
+            if (!$force) throw new TirageDejaLanceException();
+            $this->resultatRepository->deleteByOccasion($occasion);
+        }
+
+        $participants = $occasion->getParticipants();
+        $tabExclusions = array_fill(0, count($participants), []);
+
+        $exclusions = $this->exclusionRepository->readByParticipantsOccasion($occasion);
+        foreach ($exclusions as $exclusion) {
+            $idxQuiOffre = array_search($exclusion->getQuiOffre(), $participants);
+            $idxQuiNeDoitPasRecevoir = array_search($exclusion->getQuiNeDoitPasRecevoir(), $participants);
+            $tabExclusions[$idxQuiOffre][] = $idxQuiNeDoitPasRecevoir;
+        }
+
+        $resultatsPasses = $this->resultatRepository->readByParticipantsOccasion($occasion);
+        foreach ($resultatsPasses as $resultatPasse) {
+            $idxQuiOffre = array_search($resultatPasse->getQuiOffre(), $participants);
+            $idxQuiNeDoitPasRecevoir = array_search($resultatPasse->getQuiRecoit(), $participants);
+            $tabExclusions[$idxQuiOffre][] = $idxQuiNeDoitPasRecevoir;
+        }
+
+        $tabResultats = false;
+        for ($iter = 0; !$tabResultats && $iter < ($nbMaxIter | 10); $iter++) {
+            $tabResultats = $this->lanceTirage_(count($participants), $tabExclusions);
+        }
+        if (!$tabResultats) throw new TirageEchoueException();
+
+        $resultats = [];
+        foreach ($tabResultats as $idxQuiOffre => $idxQuiRecoit) {
+            $quiOffre = $participants[$idxQuiOffre];
+            $quiRecoit = $participants[$idxQuiRecoit];
+            $resultats[] = $this->resultatRepository->create($occasion, $quiOffre, $quiRecoit);
+        }
+
+        foreach ($participants as $participant) {
+            $this->mailPlugin->envoieMailTirageFait($participant, $occasion);
+        }
+
+        return $resultats;
+    }
+
+    /**
+     * @param int[][] $tabExclusions
+     * @return int[]|false
+     */
+    private function lanceTirage_(int $nbParticipants, array $tabExclusions)
+    {
+        srand();
+
+        $dejaTire = array_fill(0, $nbParticipants, false);
+        $tabResultats = array_fill(0, $nbParticipants, 0);
+
+        for ($quiOffre = 0; $quiOffre < $nbParticipants; $quiOffre++) {
+            $tabPossibilites = [];
+            for ($quiRecoit = 0; $quiRecoit < $nbParticipants; $quiRecoit++) {
+                if ($quiRecoit != $quiOffre && !$dejaTire[$quiRecoit] && !in_array($quiRecoit, $tabExclusions[$quiOffre])) {
+                    $tabPossibilites[] = $quiRecoit;
+                }
+            }
+
+            if (count($tabPossibilites) == 0) return false;
+
+            $quiRecoit = $tabPossibilites[rand(0, count($tabPossibilites) - 1)];
+            $tabResultats[$quiOffre] = $quiRecoit;
+            $dejaTire[$quiRecoit] = true;
+        }
+
+        return $tabResultats;
+    }
+
+    /**
      * @return Occasion[]
      * @throws PasAdminException quand l'utilisateur n'est pas admin
      */
     public function listeOccasions(
         Auth $auth
-    ): array
-    {
+    ): array {
         if (!$auth->estAdmin()) throw new PasAdminException();
         return $this->occasionRepository->readAll();
     }
@@ -169,8 +252,7 @@ class OccasionPort
     public function listeOccasionsParticipant(
         Auth $auth,
         Utilisateur $participant
-    ): array
-    {
+    ): array {
         if (!$auth->estUtilisateur($participant) && !$auth->estAdmin()) throw new PasUtilisateurNiAdminException();
         return $this->occasionRepository->readByParticipant($participant);
     }